Autoformat -> Make indentation 4 spaces as per group decision.
Kevin and Aaron agreed to use four spaces as indentation.
The reason for this that four spaces make it very clear and easy to see which indentation level a piece of code belongs to.
Code indentation with two spaces takes more mental capacity to distinguish levels of indentation.
We have decided against tabs for indentation because tabs are not universally defined to be a specific amount of spaces.
A tab may be 8 spaces on one machine and 6 spaces on another.